github: add GUI abstraction layer.
In preparation for adopting the capture and reply framework, the GUI static, global interface needs to abstracted into a contextual service.

 - [x] Provide `IRuntimeService` based service, encapsulating GUI API.
 - [x] Provide an extention of the `MS.ALM.Base` type exposing the new runtime service.
 - [x] Adopt contextual service provider based GUI API.
